+++ centralApi/src/database/seedContent/vaultDescriptionCobreFreeSurfer.ts
@@ -1,0 +1,15 @@
+export default `
+**Data description**: Contains 149 subjects: summary-level Volumetric measurements (text) of Brain ROI's processed via FreeSurferv5.3; no images
+
+**Example Files** (Headers Only):
+- [covariates.csv](https://git-link.vercel.app/api/download?url=https%3A%2F%2Fgithub.com%2FNeuroFlame%2Fvault-trends-freesurfer-cobre%2Fblob%2Fmain%2Fcovariates.csv)
+- [data.csv](https://git-link.vercel.app/api/download?url=https%3A%2F%2Fgithub.com%2FNeuroFlame%2Fvault-trends-freesurfer-cobre%2Fblob%2Fmain%2Fdata.csv)
+
+**Covariates**: Age (18-65), Gender (M-115, Females-37), Patients/controls(81/68), 8 cognitive variables
+
+**Scanner info**: 3T Siemens Trio MRI scanner
+
+**Data Use Agreement\***: "If you use this vault data for a reseach publication....I agree to cite this paper: \[1\] C. J. Aine, H. J. Bockholt, J. R. Bustillo, J. M. Canive, A. Caprihan, C. Gasparovic, F. M. Hanlon, J. M. Houck, R. E. Jung, J. Lauriello, J. Liu, A. R. Mayer, N. I. Perrone-Bizzozero, S. Posse, J. M. Stephen, J. A. Turner, V. P. Clark, and V. D. Calhoun, "Multimodal neuroimaging in schizophrenia: Description and dissemination," _Neuroinformatics,_ vol. 15, pp. 343–364, Oct 2017, PMC5671541."
+
+**More info:** [https://github.com/NeuroFlame/vault-trends-freesurfer-cobre](https://github.com/NeuroFlame/vault-trends-freesurfer-cobre)
+`